\name{summary}
\alias{summary}
\title{A Brief Summary of Control Chart Parameters and Data}
\description{
  This function provides a summary for the control charts generated by the
  \code{mxrpc} and \code{mxspc} functions. This function also provide a brief summary of data being used for analysis.
}
\usage{
summary(object, ...)
}
\arguments{
  \item{object}{An object of class \code{control.chart} created by either \code{mxrpc} or \code{mxspc}.}
  \item{...}{Additional arguments passed to the method (currently unused).}
}
\details{
  This method prints the control chart parameters such as subgroup number, sample size,
  control limits, summary statistics of the plotting statistic and summary statistics of the Maxwell-distributed data.
}
\value{
  This function prints a summary of the control chart parameters and data.
}
\examples{

# Assuming `t1` is an object returned by the mxspc() function
t1 <- mxspc(m = 40, n = 3, alpha = 0.0027, sigma = 1.5, limit = "PCL")
summary(t1)
}

